API/APIs/requirements.txt
absl-py==0.7.1
alabaster==0.7.12
argon2-cffi==18.1.0
asn1crypto==0.24.0
astor==0.7.1
astroid==2.1.0
Babel==2.6.0
backcall==0.1.0
bcrypt==3.1.4
bleach==3.1.0
certifi==2019.3.9
cffi==1.11.5
chardet==3.0.4
cloudpickle==0.8.0
codegen==1.0
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.6.1
cycler==0.10.0
decorator==4.4.0
defusedxml==0.5.0
dill==0.2.5
Django==2.2
django-bootstrap3==10.0.1
django-braces==1.13.0
django-debug-toolbar==1.9.1
django-estimators==0.2.1
django-rest-swagger==2.2.0
djangorestframework==3.9.2
docutils==0.14
entrypoints==0.3
Faker==0.8.15
future==0.17.1
gast==0.2.2
google-pasta==0.1.5
grpcio==1.19.0
h5py==2.9.0
idna==2.8
imagesize==1.1.0
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
isort==4.3.16
itypes==1.1.0
jedi==0.13.3
Jinja2==2.10
jsonfield==2.0.2
jsonschema==2.6.0
jupyter-client==5.2.4
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
keyring==18.0.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
Markdown==3.1
MarkupSafe==1.0
matplotlib==3.0.3
mccabe==0.6.1
mistune==0.8.3
mock==2.0.0
msgpack==0.6.1
nbconvert==5.4.1
nbformat==4.4.0
nltk==3.4
np-utils==0.5.10.0
numpy==1.16.2
numpydoc==0.8.0
openapi-codec==1.3.2
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.3.4
patterns==0.3
pbr==5.1.3
pickleshare==0.7.5
Pillow==5.1.0
prompt-toolkit==2.0.9
protobuf==3.7.1
psutil==5.4.7
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.3.1
pylint==2.2.2
pyOpenSSL==19.0.0
pyparsing==2.3.1
pyreadline==2.1
PySocks==1.6.8
python-dateutil==2.8.0
pytz==2018.4
pywin32==223
PyYAML==5.1
pyzmq==17.1.2
QtAwesome==0.5.7
qtconsole==4.4.3
QtPy==1.7.0
requests==2.21.0
rope==0.12.0
scikit-learn==0.20.3
scipy==1.2.1
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
snowballstemmer==1.2.1
Sphinx==1.8.5
sphinxcontrib-websupport==1.1.0
spyder==3.3.4
spyder-kernels==0.4.4
sqlparse==0.2.4
tb-nightly==1.14.0a20190301
tensorboard==1.13.1
tensorflow==2.0.0a0
tensorflow-estimator==1.13.0
termcolor==1.1.0
testpath==0.4.2
text-unidecode==1.2
tf-estimator-nightly==1.14.0.dev2019030115
tornado==5.1.1
traitlets==4.3.2
typed-ast==1.1.0
uritemplate==3.0.0
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.2
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.10.11
